Add from_{source} functions to grab_artwork
Instead of a single source and logic to grab artwork,
we use from_{source} to allow for more options.
This commit adds google images as a source.
The major drawback of using google is that we need an API key.
So, for other users, artwork grabber routine, which currently
does not exist, must default to from_instagram.

from_google merely makes the logic simpler.